Add files via upload
This project has been done in team of 4 people. The goal was to develop a mathematical model for the Carrier Sence Multiple Access with Collision Avoidance Protocol.  To evaluate the quality of our mathematical estimations, we had to build this simulator. 
It took a while to complete the project, but it was richly challenging and worth to work on.

As the result of the research we got as the maximal relative error (taking the simulator as stallion) of our metrics are around 8-10%, which is appreciable, but might be improved if we chose math apparatus. 

I will share the papers we published on the research project ! If interested take a look and hopefully you will join the boat:)